更改音乐数据源。

+++ Music/src/main/java/com/gfd/music/entity/MvData.kt
@@ -8,4 +8,4 @@ package com.gfd.music.entity
  */
 data class MvData(var name: String = "慕涵盛华", var id: String, var pic: String, var des: String, var autor: String,
                   var playCount: Int, var artistId: Int = 0, var score: Int = 0, var artistName:String = "",
-                  var duration:Int = 0,var videoUrl: String = "")
+                  var duration:Long = 0,var videoUrl: String = "")
