+++ Music/src/main/java/com/gfd/music/entity/Radio.kt
@@ -55,7 +55,7 @@ data class Dj(
         val backgroundImgId: Long,
         val backgroundImgIdStr: String,
         val backgroundUrl: String,
-        val birthday: Int,
+        val birthday: Long,
         val city: Int,
         val defaultAvatar: Boolean,
         val description: String,
